[nmpc][py]: move other unused file to the archive folder
