+++ aerial_robot_control/scripts/nmpc/tilt_qd/archive/tilt_qd_servo_old_cost.py
@@ -12,6 +12,7 @@ import sys
 import numpy as np
 import yaml
 import rospkg
+from tf_conversions import transformations as tf
 from acados_template import AcadosModel, AcadosOcp, AcadosOcpSolver
 import casadi as ca
 
@@ -22,19 +23,20 @@ from phys_param_beetle_art import *
 # read parameters from yaml
 rospack = rospkg.RosPack()
 
-nmpc_param_path = os.path.join(rospack.get_path("beetle"), "config", "BeetleNMPCVelInput.yaml")
+nmpc_param_path = os.path.join(rospack.get_path("beetle"), "config", "BeetleNMPCOldServoCost.yaml")
 with open(nmpc_param_path, "r") as f:
     nmpc_param_dict = yaml.load(f, Loader=yaml.FullLoader)
 nmpc_params = nmpc_param_dict["controller"]["nmpc"]
 nmpc_params["N_node"] = int(nmpc_params["T_pred"] / nmpc_params["T_integ"])
 
 
-class NMPCTiltQdServoVelInput(NMPCBase):
+class NMPCTiltQdServoOldCost(NMPCBase):  # old cost means the quadrotor one, not ac_d.
     def __init__(self):
-        super(NMPCTiltQdServoVelInput, self).__init__()
+        super(NMPCTiltQdServoOldCost, self).__init__()
+        self.t_servo = t_servo
 
     def set_name(self) -> str:
-        model_name = "tilt_qd_servo_vel_input_mdl"
+        model_name = "tilt_qd_servo_old_cost_mdl"
         return model_name
 
     def set_ts_ctrl(self) -> float:
@@ -81,8 +83,8 @@ class NMPCTiltQdServoVelInput(NMPCBase):
         a2c = ca.SX.sym("a2c")
         a3c = ca.SX.sym("a3c")
         a4c = ca.SX.sym("a4c")
-        ad_c = ca.vertcat(a1c, a2c, a3c, a4c)
-        controls = ca.vertcat(ft, ad_c)
+        ac = ca.vertcat(a1c, a2c, a3c, a4c)
+        controls = ca.vertcat(ft, ac)
 
         # transformation matrix
         row_1 = ca.horzcat(
@@ -163,7 +165,7 @@ class NMPCTiltQdServoVelInput(NMPCBase):
             (wy * qw - wz * qx + wx * qz) / 2,
             (wz * qw + wy * qx - wx * qy) / 2,
             ca.mtimes(inv_iv, (-ca.cross(w, ca.mtimes(iv, w)) + tau_u_b)),
-            ad_c,
+            (ac - a) / t_servo,
         )
 
         # function
@@ -176,7 +178,7 @@ class NMPCTiltQdServoVelInput(NMPCBase):
         qe_z = -qxr * qy + qx * qyr + qwr * qz - qw * qzr
 
         state_y = ca.vertcat(p, v, qwr, qe_x + qxr, qe_y + qyr, qe_z + qzr, w, a)
-        control_y = ca.vertcat(ft, ad_c)
+        control_y = ca.vertcat(ft, ac)  # key difference
 
         # acados model
         x_dot = ca.SX.sym("x_dot", 17)
@@ -255,10 +257,10 @@ class NMPCTiltQdServoVelInput(NMPCBase):
                 nmpc_params["Rt"],
                 nmpc_params["Rt"],
                 nmpc_params["Rt"],
-                nmpc_params["Rac_d"],
-                nmpc_params["Rac_d"],
-                nmpc_params["Rac_d"],
-                nmpc_params["Rac_d"],
+                nmpc_params["Rac"],
+                nmpc_params["Rac"],
+                nmpc_params["Rac"],
+                nmpc_params["Rac"],
             ]
         )
         print("R: \n", R)
@@ -394,7 +396,7 @@ class NMPCTiltQdServoVelInput(NMPCBase):
 
         return solver
 
-    def create_xr_ur_converter(self) -> XrUrConverterBase:
+    def create_xr_ur_converter(self):
         return XrUrConverter()
 
 
@@ -423,9 +425,64 @@ class XrUrConverter(XrUrConverterBase):
         self.alloc_mat_pinv = self._get_alloc_mat_pinv()
         self.ocp_N = nmpc_params["N_node"]
 
+    def pose_point_2_xr_ur(self, target_xyz, target_rpy):
+        roll = target_rpy.item(0)
+        pitch = target_rpy.item(1)
+        yaw = target_rpy.item(2)
+
+        q = tf.quaternion_from_euler(roll, pitch, yaw, axes="sxyz")
+        target_qwxyz = np.array([[q[3], q[0], q[1], q[2]]]).T
+
+        # convert [0,0,gravity] to body frame
+        q_inv = tf.quaternion_inverse(q)
+        rot = tf.quaternion_matrix(q_inv)
+        fg_i = np.array([0, 0, self.mass * self.gravity, 0])
+        fg_b = rot @ fg_i
+        target_wrench = np.array([[fg_b.item(0), fg_b.item(1), fg_b.item(2), 0, 0, 0]]).T
+
+        # a quicker method if alloc_mat is dynamic:  x, _, _, _ = np.linalg.lstsq(alloc_mat, target_wrench, rcond=None)
+        x = self.alloc_mat_pinv @ target_wrench
+
+        a1_ref = np.arctan2(x[0, 0], x[1, 0])
+        ft1_ref = np.sqrt(x[0, 0] ** 2 + x[1, 0] ** 2)
+        a2_ref = np.arctan2(x[2, 0], x[3, 0])
+        ft2_ref = np.sqrt(x[2, 0] ** 2 + x[3, 0] ** 2)
+        a3_ref = np.arctan2(x[4, 0], x[5, 0])
+        ft3_ref = np.sqrt(x[4, 0] ** 2 + x[5, 0] ** 2)
+        a4_ref = np.arctan2(x[6, 0], x[7, 0])
+        ft4_ref = np.sqrt(x[6, 0] ** 2 + x[7, 0] ** 2)
+
+        # get x and u, set reference
+        ocp_N = self.ocp_N
+
+        xr = np.zeros([ocp_N + 1, self.nx])
+        xr[:, 0] = target_xyz.item(0)  # x
+        xr[:, 1] = target_xyz.item(1)  # y
+        xr[:, 2] = target_xyz.item(2)  # z
+        xr[:, 6] = target_qwxyz.item(0)  # qx
+        xr[:, 7] = target_qwxyz.item(1)  # qx
+        xr[:, 8] = target_qwxyz.item(2)  # qy
+        xr[:, 9] = target_qwxyz.item(3)  # qz
+        xr[:, 13] = a1_ref
+        xr[:, 14] = a2_ref
+        xr[:, 15] = a3_ref
+        xr[:, 16] = a4_ref
+
+        ur = np.zeros([ocp_N, self.nu])
+        ur[:, 0] = ft1_ref
+        ur[:, 1] = ft2_ref
+        ur[:, 2] = ft3_ref
+        ur[:, 3] = ft4_ref
+        ur[:, 4] = a1_ref
+        ur[:, 5] = a2_ref
+        ur[:, 6] = a3_ref
+        ur[:, 7] = a4_ref
+
+        return xr, ur
+
 
 if __name__ == "__main__":
-    nmpc = NMPCTiltQdServoVelInput()
+    nmpc = NMPCTiltQdServoOldCost()
 
     acados_ocp_solver = nmpc.get_ocp_solver()
     print("Successfully initialized acados ocp: ", acados_ocp_solver.acados_ocp)
